+++ tasks/modules/utils.js
@@ -11,6 +11,40 @@ function makeRelativePath(folderpath, filename) {
 }
 exports.makeRelativePath = makeRelativePath;
 
+// Finds the longest common section of a collection of strings.
+// Simply sorting and comparing first and last http://stackoverflow.com/a/1917041/390330
+function sharedStart(array) {
+    if (array.length === 0) {
+        throw 'Cannot find common root of empty array.';
+    }
+    var A = array.slice(0).sort(), firstWord = A[0], lastWord = A[A.length - 1];
+
+    if (firstWord === lastWord) {
+        return firstWord;
+    } else {
+        var i = -1;
+        do {
+            i += 1;
+            var firstWordChar = firstWord.charAt(i);
+            var lastWordChar = lastWord.charAt(i);
+        } while(firstWordChar === lastWordChar);
+
+        return firstWord.substring(0, i);
+    }
+}
+
+// Finds the common system path between paths
+// Explanation of how is inline
+function findCommonPath(paths, pathSeperator) {
+    // Now for "C:\u\starter" "C:\u\started" => "C:\u\starte"
+    var largetStartSegement = sharedStart(paths);
+
+    // For "C:\u\starte" => C:\u\
+    var ending = largetStartSegement.lastIndexOf(pathSeperator);
+    return largetStartSegement.substr(0, ending);
+}
+exports.findCommonPath = findCommonPath;
+
 /**
 * Returns the result of an array inserted into another, starting at the given index.
 */