+++ tasks/ts.ts
@@ -182,6 +182,15 @@ function pluginFn(grunt: IGrunt) {
                 return path.resolve(filename) === outFile_d_ts;
             }
 
+            // see https://github.com/grunt-ts/grunt-ts/issues/77
+            function isBaseDirFile(filename: string, targetFiles: string[]) {
+                var baseDirFile: string = 'ignoreBaseDirFile.ts';
+                if (!target.baseDir) {
+                    target.baseDir = utils.findCommonPath(targetFiles, '/');
+                }
+                return path.resolve(filename) === path.resolve(path.join(target.baseDir, baseDirFile));
+            }
+
             // Create an amd loader?
             var amdloader = target.amdloader;
             var amdloaderFile;
@@ -236,18 +245,18 @@ function pluginFn(grunt: IGrunt) {
                     return !stats.isDirectory();
                 });
 
-                // Clear the files of output.d.ts and reference.ts
+                // Clear the files of output.d.ts and reference.ts and baseDirFile
                 files = _.filter(files, (filename) => {
-                    return (!isReferenceFile(filename) && !isOutFile(filename));
+                    return (!isReferenceFile(filename) && !isOutFile(filename) && !isBaseDirFile(filename, files));
                 });
 
                 ///// Html files:
                 // Note:
                 //    compile html files must be before reference file creation                
-                var generatedHtmlFiles = [];
+                var generatedFiles = [];
                 if (currenttask.data.html) {
                     var htmlFiles = grunt.file.expand(currenttask.data.html);
-                    generatedHtmlFiles = _.map(htmlFiles, (filename) => html2tsModule.compileHTML(filename));
+                    generatedFiles = _.map(htmlFiles, (filename) => html2tsModule.compileHTML(filename));
                 }
 
                 ///// Template cache
@@ -279,7 +288,7 @@ function pluginFn(grunt: IGrunt) {
                 // Create a reference file if specified
                 if (!!referencePath) {
                     var result = timeIt(() => {
-                        return referenceModule.updateReferenceFile(files, generatedHtmlFiles, referenceFile, referencePath);
+                        return referenceModule.updateReferenceFile(files, generatedFiles, referenceFile, referencePath);
                     });
                     if (result.it === true) {
                         grunt.log.writeln(('Updated reference file (' + result.time + 'ms).').green);
@@ -290,7 +299,7 @@ function pluginFn(grunt: IGrunt) {
                 // Create the amdLoader if specified 
                 if (!!amdloaderPath) {
                     var referenceOrder: amdLoaderModule.IReferences
-                        = amdLoaderModule.getReferencesInOrder(referenceFile, referencePath, generatedHtmlFiles);
+                        = amdLoaderModule.getReferencesInOrder(referenceFile, referencePath, generatedFiles);
                     amdLoaderModule.updateAmdLoader(referenceFile, referenceOrder, amdloaderFile, amdloaderPath, target.outDir);
                 }